Updated token models.

+++ CADCodeProxy/Machining/Tokens/FreePocketArcSegment.cs
@@ -22,12 +22,12 @@ public record FreePocketArcSegment : IRoutingToken, IMachiningOperation {
     void IMachiningOperation.AddToCode(CADCodeCodeClass code, double xOffset, double yOffset) {
 
         code.DefinePocket(
-            StartX: (float)Start.X + (float)xOffset,
-            StartY: (float)Start.Y + (float)yOffset,
-            StartZ: (float)StartDepth,
-            EndX: (float)End.X + (float)xOffset,
-            EndY: (float)End.Y + (float)yOffset,
-            Endz: (float)EndDepth,
+            StartX: (float) (Start.X + xOffset),
+            StartY: (float) (Start.Y + yOffset),
+            StartZ: (float) StartDepth,
+            EndX: (float) (End.X + xOffset),
+            EndY: (float) (End.Y + yOffset),
+            Endz: (float) EndDepth,
             CenterX: 0,
             CenterY: 0,
             CenterZ: 0,
